+++ wasm/src/data.rs
@@ -1,13 +1,17 @@
 use std::mem::ManuallyDrop;
+use std::ops::Deref;
 use std::pin::Pin;
+use std::ptr::NonNull;
 
 use blackbox_log::data::Stats;
 use blackbox_log::frame::{Frame, GpsFrame, MainFrame, SlowFrame};
 use blackbox_log::prelude::*;
 use blackbox_log::units::{si, Time};
-use blackbox_log::Reader;
+use blackbox_log::{FieldFilterSet, Reader};
 
-use crate::headers::WasmHeaders;
+use crate::headers::{WasmFrameDef, WasmHeaders};
+use crate::owned_slice::AllocError;
+use crate::str::WasmStr;
 use crate::{IntoWasmFfi, OwnedSlice, Shared, WasmByValue, WasmFfi};
 
 // SAFETY: field order *must* be `parser` first, then `headers`, then `data` to
@@ -26,17 +30,35 @@ impl WasmDataParser {
         headers: Shared<Headers<'static>>,
         reader: Reader<'static>,
         data: Shared<OwnedSlice<u8>>,
+        filters: Option<Box<WasmFieldFilterSetBuilder>>,
     ) -> Self {
         // SAFETY: this is only used to create the `DataParser`, which is guaranteed to
         // be dropped before `headers` by the declaration order in the struct
         let headers_ref = unsafe { headers.deref_static() };
 
+        let filters = filters.map(|filters| filters.build()).unwrap_or_default();
+
         Self {
             parsed: Box::pin(WasmParseEvent::default()),
-            parser: DataParser::new(reader, headers_ref),
+            parser: DataParser::with_filters(reader, headers_ref, &filters),
             _headers: headers,
             _data: data,
         }
+    }
+
+    fn main_def(&self) -> WasmFrameDef {
+        WasmFrameDef::from(&self.parser.main_frame_def())
+    }
+
+    fn slow_def(&self) -> WasmFrameDef {
+        WasmFrameDef::from(&self.parser.slow_frame_def())
+    }
+
+    fn gps_def(&self) -> WasmFrameDef {
+        self.parser
+            .gps_frame_def()
+            .as_ref()
+            .map_or_else(WasmFrameDef::default, WasmFrameDef::from)
     }
 
     fn result_ptr(&self) -> *const WasmParseEvent {
@@ -51,6 +73,45 @@ impl WasmDataParser {
     fn next(&mut self) {
         let parsed = self.parser.next();
         *self.parsed = parsed.into();
+    }
+}
+
+#[allow(dead_code)]
+pub struct WasmFieldFilterSetBuilder {
+    main: Option<Vec<&'static str>>,
+    slow: Option<Vec<&'static str>>,
+    gps: Option<Vec<&'static str>>,
+    arena: NonNull<u8>,
+}
+
+impl_boxed_wasm_ffi!(WasmFieldFilterSetBuilder);
+
+impl WasmFieldFilterSetBuilder {
+    fn new(arena_size: usize, main: isize, slow: isize, gps: isize) -> Result<Self, AllocError> {
+        Ok(Self {
+            main: usize::try_from(main).ok().map(Vec::with_capacity),
+            slow: usize::try_from(slow).ok().map(Vec::with_capacity),
+            gps: usize::try_from(gps).ok().map(Vec::with_capacity),
+            arena: OwnedSlice::<u8>::alloc(arena_size)?,
+        })
+    }
+
+    /// # SAFETY
+    ///
+    /// `fields` must come from a `WasmFieldFilterSetBuilder` and `str` must be
+    /// allocated in the same struct's `arena`.
+    unsafe fn push(fields: &mut Option<Vec<&'static str>>, str: WasmStr) {
+        let str: &str = str.deref();
+        let str: &'static str = std::mem::transmute(str);
+        fields.as_mut().unwrap().push(str);
+    }
+
+    fn build(self) -> FieldFilterSet {
+        FieldFilterSet {
+            main: self.main.as_deref().map(Into::into),
+            slow: self.slow.as_deref().map(Into::into),
+            gps: self.gps.as_deref().map(Into::into),
+        }
     }
 }
 
@@ -252,12 +313,42 @@ where
 {
 }
 
+#[repr(C)]
+struct FieldFilterSetBuilderNew(<Box<WasmFieldFilterSetBuilder> as WasmFfi>::Ffi, *const u8);
+
+// SAFETY: repr(C) & where bounds for each field
+unsafe impl WasmByValue for FieldFilterSetBuilderNew
+where
+    <Box<WasmFieldFilterSetBuilder> as WasmFfi>::Ffi: WasmByValue,
+    *const u8: WasmByValue,
+{
+}
+
 wasm_export!(free data_free: Box<WasmDataParser>);
 wasm_export! {
-    fn data_new(headers: ref Box<WasmHeaders>) -> DataNew {
-        let data = Box::new(headers.get_data_parser());
+    fn data_new(headers: ref Box<WasmHeaders>, filters: owned *mut WasmFieldFilterSetBuilder) -> DataNew {
+        let filters = if filters.is_null() {
+            None
+        } else {
+            // SAFETY: `filters` is from `Box::into_raw` in `filter_new`
+            Some(unsafe { Box::from_raw(filters) })
+        };
+
+        let data = Box::new(headers.get_data_parser(filters));
         let result = data.result_ptr();
         DataNew(data.into_ffi(),result.into_ffi())
+    }
+
+    fn data_mainDef(parser: ref Box<WasmDataParser>) -> Box<WasmFrameDef> {
+        Box::new(parser.main_def())
+    }
+
+    fn data_slowDef(parser: ref Box<WasmDataParser>) -> Box<WasmFrameDef> {
+        Box::new(parser.slow_def())
+    }
+
+    fn data_gpsDef(parser: ref Box<WasmDataParser>) -> Box<WasmFrameDef> {
+        Box::new(parser.gps_def())
     }
 
     fn data_stats(parser: ref Box<WasmDataParser>) -> WasmDataStats {
@@ -275,5 +366,33 @@ wasm_export! {
 
     fn data_next(parser: ref_mut Box<WasmDataParser>) {
         parser.next();
+    }
+
+    fn filter_new(
+        arena_size: owned usize,
+        main: owned isize,
+        slow: owned isize,
+        gps: owned isize,
+    ) -> FieldFilterSetBuilderNew {
+        let builder = Box::new(WasmFieldFilterSetBuilder::new(
+            arena_size,
+            main,
+            slow,
+            gps,
+        ).unwrap());
+        let ptr = builder.arena.as_ptr();
+        FieldFilterSetBuilderNew(Box::into_raw(builder), ptr)
+    }
+
+    fn filter_main(builder: ref_mut Box<WasmFieldFilterSetBuilder>, str: owned WasmStr) {
+        WasmFieldFilterSetBuilder::push(&mut builder.main, str);
+    }
+
+    fn filter_slow(builder: ref_mut Box<WasmFieldFilterSetBuilder>, str: owned WasmStr) {
+        WasmFieldFilterSetBuilder::push(&mut builder.slow, str);
+    }
+
+    fn filter_gps(builder: ref_mut Box<WasmFieldFilterSetBuilder>, str: owned WasmStr) {
+        WasmFieldFilterSetBuilder::push(&mut builder.gps, str);
     }
 }